feat: async get loro
