+++ packages/blocky-example/app/app.scss
@@ -111,7 +111,7 @@ body {
 
   p {
     margin-left: 10px;
-    font-size: 13px;
+    font-size: 12px;
     font-family: system-ui, -apple-system, BlinkMacSystemFont, 'Segoe UI', Roboto, Oxygen, Ubuntu, Cantarell, 'Open Sans', 'Helvetica Neue', sans-serif;
     color: var(--primary-text-color);
   }
